Added string type, comments, source loading, stdlib
Added string type, still need to change a few builtins to manipulate
them properly.

Added comments, denoted by a ;.

Source loading, the Nitrogen executable can be called with a filename
as an argument in which case it will execute the code in the file.
Also added 'load' keyword which will load in the source file with the
given name.

Created standard library with basic functions such as min, max,
fibinocci, list manipulations, fun declarations with 'fun' keyword,
logical comparators not, or, and.