Updated stdlib with new protected definitions
Changed all definitions to protected. Also created a protected functions
definition that effectivally makes user-defined functions act as builtins
in the fact that they can't be changed. Ideally the pfun functions will
primarily be used in the standard library and other first-party libraries.
But it will be accessable to users.